+++ src/core/index.ts
@@ -2,7 +2,6 @@ export * from "./nodes/abstract-node"
 export * from "./nodes/complex-node"
 export * from "./nodes/immutable-node"
 export * from "./nodes/reference-node"
-export * from "./mst-node"
 export * from "./action"
 export * from "./json-patch"
 export * from "./mst-operations"
