+++ src/types/utility-types/optional.ts
@@ -1,5 +1,6 @@
 import {Type, IType, TypeFlags} from "../type"
 import { IContext, IValidationResult, typecheck, typeCheckSuccess, typeCheckFailure } from "../type-checker"
+import { isMST, getMSTAdministration, ComplexNode, AbstractNode  } from "../../core"
 
 export type IFunctionReturn<T> = () => T
 export type IOptionalValue<S, T> = S | T | IFunctionReturn<S> | IFunctionReturn<T>
@@ -58,5 +59,4 @@ export function optional<S, T>(type: IType<S, T>, defaultValueOrFunction: any): 
     return new OptionalValue(type, defaultValueOrFunction)
 }
 
-import { isMST, getMSTAdministration, ComplexNode, AbstractNode  } from "../../core"
